restructured files to reduce runtime in case of large number of parsing items. now properly using both friendly names and section data plugins
